moved shell_body and remove_faces to body.py and tests to test_design.py per Roberto's comments and scrum discussion today
