+++ src/ansys/geometry/core/designer/body.py
@@ -53,6 +53,8 @@ from ansys.api.geometry.v0.commands_pb2 import (
     AssignMidSurfaceThicknessRequest,
     ImprintCurvesRequest,
     ProjectCurvesRequest,
+    ShellRequest,
+    RemoveFacesRequest
 )
 from ansys.api.geometry.v0.commands_pb2_grpc import CommandsStub
 from ansys.geometry.core.connection.client import GrpcClient
@@ -76,6 +78,7 @@ from ansys.geometry.core.math.point import Point3D
 from ansys.geometry.core.math.vector import UnitVector3D
 from ansys.geometry.core.misc.checks import (
     check_type,
+    check_type_all_elements_in_iterable,
     ensure_design_is_active,
     min_backend_version,
 )
@@ -552,6 +555,38 @@ class IBody(ABC):
         """
         return
 
+    @abstractmethod
+    def shell_body(self, offset: Real) -> bool:
+        """Shell the body to the thickness specified.
+
+        Parameters
+        ----------
+        offset : Real
+            Shell thickness.
+
+        Returns
+        -------
+        bool
+            ``True`` when successful, ``False`` when failed.
+        """
+        return
+    
+    @abstractmethod
+    def remove_faces(self, selection: Union["Face", list["Face"]], offset: Real) -> bool:
+        """Shell by removing a given set of faces.
+        
+        Parameters
+        ----------
+        selection : Face | List[Face]
+            Face or faces to be removed.
+        offset : Real
+            Shell thickness.
+        Returns
+        -------
+        bool
+            ``True`` when successful, ``False`` when failed.
+        """
+    
     @abstractmethod
     def plot(
         self,
@@ -1162,6 +1197,40 @@ class MasterBody(IBody):
         else:
             return comp
 
+    @protect_grpc
+    def shell_body(self, offset: Real) -> bool:  # noqa: D102
+        check_type(offset, Real)
+
+        self._grpc_client.log.debug(f"Shelling body {self.id} to offset {offset}.")
+
+        result = self._commands_stub.Shell(
+            ShellRequest(
+                selection=self._grpc_id,
+                offset=offset,
+            )
+        )
+
+        return result.success
+    
+    @protect_grpc
+    @reset_tessellation_cache
+    def remove_faces(self, selection: Union["Face", list["Face"]], offset: Real) -> bool:  # noqa: D102
+        selection: list[Face] = selection if isinstance(selection, list) else [selection]
+        
+        check_type(offset, Real)
+        check_type_all_elements_in_iterable(selection, Face)
+
+        self._grpc_client.log.debug(f"Removing faces to shell body {self.id}.")
+
+        result = self._commands_stub.RemoveFaces(
+            RemoveFacesRequest(
+                selection=[face._grpc_id for face in selection],
+                offset=offset,
+            )
+        )
+
+        return result.success
+
     def plot(  # noqa: D102
         self,
         merge: bool = True,
@@ -1578,6 +1647,14 @@ class Body(IBody):
         self, merge: bool = False
     ) -> Union["PolyData", "MultiBlock"]:
         return self._template.tessellate(merge, self.parent_component.get_world_transform())
+
+    @ensure_design_is_active
+    def shell_body(self, offset: Real) -> bool:
+        return self._template.shell_body(offset)
+    
+    @ensure_design_is_active
+    def remove_faces(self, selection: Union["Face", list["Face"]], offset: Real) -> bool:  # noqa: D102
+        return self._template.remove_faces(selection, offset)
 
     def plot(  # noqa: D102
         self,
